Add docker samples. Make some refactoring
