+++ Dockerfile
@@ -1,5 +1,11 @@
 FROM python:3.7-alpine
 
-COPY requirements.txt /
 RUN apk add --update --no-cache g++ gcc libxslt-dev
+
+COPY requirements.txt /
+
 RUN pip install -r requirements.txt
+
+WORKDIR /home/manager-ai
+
+CMD ["python", "./src/main.py"]
