+++ docker-compose.yml
@@ -1,11 +1,17 @@
 version: "3"
 services:
-  m-ai:
+  managerai:
     build: .
-    command: python src # TODO: fix
-    links:
+    volumes:
+      - .:/home/manager-ai/
+      - ./config.docker.yml:/home/manager-ai/config.yml
+    depends_on:
       - mongodb
+
   mongodb:
     image: mongo:4.2
     ports:
       - 27017:27017
+
+volumes:
+  mongodata:
