+++ src/tasks/hello_world.py
@@ -1,5 +1,5 @@
-from src.settings import logger
-from src.tasks.base import BaseTask
+from settings import logger
+from tasks.base import BaseTask
 
 
 class HelloWorldTask(BaseTask):
